feat: better error when package is not found
When a package is not found during installation, show which package is requiring it (if any) in the
error message. If there is no parent request, throw the same error as
before.
Add isMessageError boolean value to MessageError class to distinguish
between errors thrown by yarn from errors thrown by yarn dependencies.